// src/app/hero/hero-actions.enum.ts
export enum heroActions {
  CREATE = '[HERO] CREATE',
  CREATE_SUCCESS = '[HERO] CREATE_SUCCESS',
  CREATE_ERROR = '[HERO] CREATE_ERROR',
  DEKETE = '[HERO] DELETE',
  UPDATE = '[HERO] UPDATE',
  UPDATE_SUCCESS = '[HERO] UPDATE_SUCCESS',
  UPDATE_ERROR = '[HERO] UPDATE_ERROR',
  UPDATE_LIST = '[HERO] UPDATE_LIST',
  UPDATE_LIST_SUCCESS = '[HERO] UPDATE_LIST_SUCCESS',
  UPDATE_LIST_ERROR = '[HERO] UPDATE_LIST_ERROR',

  DELETE = '[HERO] DELETE',
  DELETE_SUCCESS = '[HERO] DELETE_SUCCESS',
  DELETE_ERROR = '[HERO] DELETE_ERROR',
}